web-animations/animation-player/animation-player-constructor.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationPlayer constructor tests</title>
<meta name="assert" content="Creates a new AnimationPlayer object">
<link rel="help" href="http://w3c.github.io/web-animations/#the-animationplayer-interface">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="author" title="Aleksei Yu. Semenov" href="mailto:a.semenov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var player = new AnimationPlayer();

    assert_equals(player.source, null, 'If constructor source parameter is ommitted ' +
        'then source attribute of newly created player should be null');
    assert_equals(player.timeline, null, 'If constructor timeline parameter is ommitted ' +
        'then timeline attribute of newly created player should be null');
}, 'The AnimationPlayer constructor without parameters specified creates ' +
    'AnimationPlayer instance with null source and null timeline');

test(function() {
    var player = new AnimationPlayer(null);

    assert_equals(player.source, null, 'If constructor source parameter is null ' +
        'then source attribute of newly created player should be null');
    assert_equals(player.timeline, null, 'If constructor timeline parameter is ommitted ' +
        'then timeline attribute of newly created player should be null');
}, 'The AnimationPlayer constructor with null source and no timeline specified creates ' +
    'AnimationPlayer instance with null source and null timeline');

test(function() {
    var player = new AnimationPlayer(null, null);

    assert_equals(player.source, null, 'If constructor source parameter is null ' +
        'then source attribute of newly created player should be null');
    assert_equals(player.timeline, null, 'If constructor timeline parameter is null ' +
        'then timeline attribute of newly created player should be null');
}, 'The AnimationPlayer constructor with null source and timeline parameters creates ' +
    'AnimationPlayer instance with null source and null timeline');

test(function() {
    var sources = [newAnimation(createDiv(this)), new AnimationGroup([]), new AnimationSequence([])];
    sources.forEach(function(source) {
        var player = new AnimationPlayer(source);

        assert_equals(player.source, source, 'Wrong source for source type ' + type(source));
    });
}, 'The AnimationPlayer constructor with source parameter specified creates ' +
    'AnimationPlayer instance with source provided');

test(function() {
    var timeline = document.timeline;
    var player = new AnimationPlayer(null, timeline);

    assert_equals(player.timeline, timeline, 'Timeline of newly created player ' +
        'should be specified by constructor timeline parameter');
}, 'The AnimationPlayer constructor with timeline parameter specified creates ' +
    'AnimationPlayer instance with timeline provided');

test(function() {
    var parent = new AnimationGroup([newAnimation(createDiv(this)),
        new AnimationGroup([]), new AnimationSequence([])]);
    parent.children.forEach(function(source) {
        var player = new AnimationPlayer(source);

        assert_equals(player.source, source, 'Source of newly created player ' +
            'should be specified by constructor source parameter');
        assert_equals(source.parent, null, 'Source should be removed from its parent ' +
            'animation group');
    });
    assert_array_equals(parent.children, [], 'Source should be removed from the parent ' +
        'group\'s children');
}, 'If source parameter of the AnimationPlayer constructor has a parent group ' +
    'then it should be removed from the group');
</script>
</body>
